# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "GoLang - Gin - RESTful - MongoDB - Redis",
        "title": "GoodBlast API",
        "termsOfService": "https://swagger.io/terms/",
        "contact": {
            "name": "Muhammet Kerem Ozturk",
            "url": "https://github.com/kerem-ozt",
            "email": "mkeremozt@gmail.com"
        },
        "license": {
            "name": "MIT License",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:3002",
    "basePath": "/",
    "paths": {
...
        "/tournament/getall": {
            "get": {
                "description": "get all tournaments",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tournament"
                ],
                "summary": "Get All Tournaments",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Page",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        },
                        "examples": {
                            "application/json": {
                                "data": {
                                    "tournaments": [
                                        {
                                            "id": "5f9b9b5b9c9b7a1b9c9b7a1b",
                                            "participants": [
                                                "5f9b9b5b9c9b7a1b9c9b7a1b"
                                            ],
                                            "minLevels": 1,
                                            "entryFee": 500,
                                            "maxParticipants": 35,
                                            "scores": [],
                                            "groups": [],
                                            "startTime": "2020-10-30T00:00:00Z",
                                            "endTime": "2020-10-31T00:00:00Z",
                                            "createdAt": "2020-10-30T00:00:00Z",
                                            "updatedAt": "2020-10-30T00:00:00Z"
                                        }
                                    ],
                                    "prev": false,
                                    "next": false
                                },
                                "message": "OK",
                                "success": true
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        },
                        "examples": {
                            "application/json": {
                                "data": null,
                                "message": "Bad Request",
                                "success": false
                            }
                        }
                    }
                }
            }
        },
        "/tournament/getbyid": {
            "get": {
                "description": "get tournament by id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tournament"
                ],
                "summary": "Get Tournament By Id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Id",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        },
                        "examples": {
                            "application/json": {
                                "data": {
                                    "tournament": {
                                        "id": "5f9b9b5b9c9b7a1b9c9b7a1b",
                                        "participants": [
                                            "5f9b9b5b9c9b7a1b9c9b7a1b"
                                        ],
                                        "minLevels": 1,
                                        "entryFee": 500,
                                        "maxParticipants": 35,
                                        "scores": [],
                                        "groups": [],
                                        "startTime": "2020-10-30T00:00:00Z",
                                        "endTime": "2020-10-31T00:00:00Z",
                                        "createdAt": "2020-10-30T00:00:00Z",
                                        "updatedAt": "2020-10-30T00:00:00Z"
                                    }
                                },
                                "message": "OK",
                                "success": true
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        },
                        "examples": {
                            "application/json": {
                                "data": null,
                                "message": "Bad Request",
                                "success": false
                            }
                        }
                    }
                }
            }
        },
        "/tournament/create": {
            "post": {
                "description": "create a new tournament",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tournament"
                ],
                "summary": "Create Tournament",
                "parameters": [
                    {
                        "description": "Tournament Request",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.TournamentRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        },
                        "examples": {
                            "application/json": {
                                "data": {
                                    "tournament": {
                                        "id": "5f9b9b5b9c9b7a1b9c9b7a1b",
                                        "participants": [
                                            "5f9b9b5b9c9b7a1b9c9b7a1b"
                                        ],
                                        "minLevels": 1,
                                        "entryFee": 500,
                                        "maxParticipants": 35,
                                        "scores": [],
                                        "groups": [],
                                        "startTime": "2020-10-30T00:00:00Z",
                                        "endTime": "2020-10-31T00:00:00Z",
                                        "createdAt": "2020-10-30T00:00:00Z",
                                        "updatedAt": "2020-10-30T00:00:00Z"
                                    }
                                },
                                "message": "OK",
                                "success": true
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        },
                        "examples": {
                            "application/json": {
                                "data": null,
                                "message": "Bad Request",
                                "success": false
                            }
                        }
                    }
                }
            }
        },
        "/tournament/creategroup": {
            "post": {
                "description": "create a new tournament group",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tournament"
                ],
                "summary": "Create Tournament Group",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        },
                        "examples": {
                            "application/json": {
                                "data": {
                                    "tournament": {
                                        "id": "5f9b9b5b9c9b7a1b9c9b7a1b",
                                        "participants": [
                                            "5f9b9b5b9c9b7a1b9c9b7a1b"
                                        ],
                                        "minLevels": 1,
                                        "entryFee": 500,
                                        "maxParticipants": 35,
                                        "scores": [],
                                        "groups": [
                                            {
                                                "groupId": "5f9b9b5b9c9b7a1b9c9b7a1b",
                                                "participants": [
                                                    "5f9b9b5b9c9b7a1b9c9b7a1b"
                                                ]
                                            }
                                        ],
                                        "startTime": "2020-10-30T00:00:00Z",
                                        "endTime": "2020-10-31T00:00:00Z",
                                        "createdAt": "2020-10-30T00:00:00Z",
                                        "updatedAt": "2020-10-30T00:00:00Z"
                                    }
                                },
                                "message": "OK",
                                "success": true
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        },
                        "examples": {
                            "application/json": {
                                "data": null,
                                "message": "Bad Request",
                                "success": false
                            }
                        }
                    }
                }
            }
        },
        "/tournament/progress": {
            "post": {
                "description": "update tournament progress",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tournament"
                ],
                "summary": "Update Tournament Progress",
                "parameters": [
                    {
                        "description": "Tournament Progress Request",
                        "name": "tournamentID",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        },
                        "examples": {
                            "application/json": {
                                "data": {
                                    "tournament": {
                                        "id": "5f9b9b5b9c9b7a1b9c9b7a1b",
                                        "participants": [
                                            "5f9b9b5b9c9b7a1b9c9b7a1b"
                                        ],
                                        "minLevels": 1,
                                        "entryFee": 500,
                                        "maxParticipants": 35,
                                        "scores": [
                                            {
                                                "userId": "5f9b9b5b9c9b7a1b9c9b7a1b",
                                                "score": 100
                                            }
                                        ],
                                        "groups": [
                                            {
                                                "groupId": "5f9b9b5b9c9b7a1b9c9b7a1b",
                                                "participants": [
                                                    "5f9b9b5b9c9b7a1b9c9b7a1b"
                                                ]
                                            }
                                        ],
                                        "startTime": "2020-10-30T00:00:00Z",
                                        "endTime": "2020-10-31T00:00:00Z",
                                        "createdAt": "2020-10-30T00:00:00Z",
                                        "updatedAt": "2020-10-30T00:00:00Z"
                                    }
                                },
                                "message": "OK",
                                "success": true
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        },
                        "examples": {
                            "application/json": {
                                "data": null,
                                "message": "Bad Request",
                                "success": false
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.Response": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Bearer-Token",
            "in": "header"
        }
    }
}